Replace yeoman boilerplate app with blank city page
The structure of this app differs a bit from previous angular
apps. Instead of a scripts dir full of random modules,
there are two dirs, `components` and `pages`. Components are
pieces of functionality that are sprinkled throughout the app,
while pages are views in the application that can be directly
routed to. Components, then, are generally a directive along
with any necessary child directives and/or services required
for it to function, and pages are a controller and partial
that is composed mostly of other components, along with a
$state definition for routing.